cdk/my-cdk/requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_s3
aws-cdk.aws_iam
aws-cdk.aws_autoscaling
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_ssm
aws-cdk.aws_secretsmanager